backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
psycopg2-binary
python-dotenv==1.0.0
numpy>=1.26,<2.0
pandas>=2.1.0
bcrypt==4.0.1
PyJWT==2.8.0
setuptools==68.0.0